Consolidate TRNTYPE lists
The acceptable values for `<STMTTRN><TRNTYPE>` are the same regardless
of whether it is a child of `<BANKTRANLIST>` or `<INVBANKTRAN>`.
See section 11.4.4 of the [OFX spec v2.3](https://financialdataexchange.org/common/Uploaded%20files/OFX%20files/OFX%20Banking%20Specification%20v2.3.pdf).
There is no reason to have two separate lists (one of which was
incomplete), so just use one.